move simple-adapter reference engine to shared code
