+++ packages/zuora-billing-adapter/src/filters/field_references.ts
@@ -13,138 +13,38 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import {
-  Element, isInstanceElement, Value, Values,
-  ReferenceExpression, InstanceElement, Field,
-} from '@salto-io/adapter-api'
-import { TransformFunc, transformValues } from '@salto-io/adapter-utils'
-import _ from 'lodash'
-import { logger } from '@salto-io/logging'
-import { values as lowerDashValues, collections } from '@salto-io/lowerdash'
+import { Element } from '@salto-io/adapter-api'
+import { references as referenceUtils } from '@salto-io/adapter-components'
+import { WORKFLOW_TYPE, TASK_TYPE, SETTINGS_TYPE_PREFIX } from '../constants'
 import { FilterCreator } from '../filter'
-import {
-  ReferenceSerializationStrategy, ExtendedReferenceTargetDefinition, ReferenceResolverFinder,
-  generateReferenceResolverFinder, FieldReferenceDefinition,
-} from '../reference_mapping'
 
-const log = logger(module)
-const { isDefined } = lowerDashValues
-const { awu } = collections.asynciterable
-type ElemLookupMapping = Record<string, Record<string, Element>>
-
-const replaceReferenceValues = async (
-  instance: InstanceElement,
-  resolverFinder: ReferenceResolverFinder,
-  elemLookupMaps: ElemLookupMapping[],
-  fieldsWithResolvedReferences: Set<string>,
-): Promise<Values> => {
-  const getRefElem = (
-    val: string | number, target: ExtendedReferenceTargetDefinition,
-  ): Element | undefined => {
-    const findElem = (value: string, targetType?: string): Element | undefined => (
-      targetType !== undefined
-        // TODO make the field we're using to look up more explicit
-        ? elemLookupMaps.map(lookup => lookup[targetType]?.[value]).find(isDefined)
-        : undefined
-    )
-
-    const elemParent = target.parent
-    const elemType = target.type
-    return findElem(
-      target.lookup(val, elemParent),
-      elemType,
-    )
-  }
-
-  const replacePrimitive = async (val: string | number, field: Field): Promise<Value> => {
-    const toValidatedReference = async (
-      serializer: ReferenceSerializationStrategy,
-      elem: Element | undefined,
-    ): Promise<ReferenceExpression | undefined> => {
-      if (elem === undefined) {
-        return undefined
-      }
-      const res = (await serializer.serialize({
-        ref: new ReferenceExpression(elem.elemID, elem),
-        field,
-      }) === val) ? new ReferenceExpression(elem.elemID) : undefined
-      if (res !== undefined) {
-        fieldsWithResolvedReferences.add(field.elemID.getFullName())
-      }
-      return res
-    }
-
-    const reference = (await awu(resolverFinder(field))
-      .filter(refResolver => refResolver.target !== undefined)
-      .map(async refResolver => toValidatedReference(
-        refResolver.serializationStrategy,
-        getRefElem(val, refResolver.target as ExtendedReferenceTargetDefinition),
-      ))
-      .filter(isDefined)
-      .toArray())
-      .pop()
-
-    return reference ?? val
-  }
-
-  const transformPrimitive: TransformFunc = ({ value, field }) => (
-    (
-      !_.isUndefined(field)
-      && (_.isString(value) || _.isNumber(value))
-    )
-      ? replacePrimitive(value, field)
-      : value
-  )
-
-  return await transformValues(
-    {
-      values: instance.value,
-      type: await instance.getType(),
-      transformFunc: transformPrimitive,
-      strict: false,
-      pathID: instance.elemID,
-    }
-  ) || instance.value
-}
-
-const mapFieldToElem = (
-  instances: InstanceElement[], fieldName: string,
-): Record<string, Element> => (
-  _(instances)
-    .filter(e => e.value[fieldName] !== undefined)
-    .map(e => [e.value[fieldName], e])
-    .fromPairs()
-    .value()
-)
-
-const groupByTypeAndField = (
-  instances: InstanceElement[], fieldName: string,
-): ElemLookupMapping => (
-  _(instances)
-    .groupBy(e => e.refType.elemID.name)
-    .mapValues(insts => mapFieldToElem(insts, fieldName))
-    .value()
-)
-
-export const addReferences = async (
-  elements: Element[],
-  defs?: FieldReferenceDefinition[]
-): Promise<void> => {
-  const resolverFinder = generateReferenceResolverFinder(defs)
-  const instances = elements.filter(isInstanceElement)
-  const elemIDLookup = groupByTypeAndField(instances, 'id')
-  const elemNameLookup = groupByTypeAndField(instances, 'name')
-  const fieldsWithResolvedReferences = new Set<string>()
-  await awu(instances).forEach(async instance => {
-    instance.value = await replaceReferenceValues(
-      instance,
-      resolverFinder,
-      [elemIDLookup, elemNameLookup],
-      fieldsWithResolvedReferences,
-    )
-  })
-  log.debug('added references in the following fields: %s', [...fieldsWithResolvedReferences])
-}
+const fieldNameToTypeMappingDefs: referenceUtils.FieldReferenceDefinition[] = [
+  {
+    src: { field: 'source_workflow_id', parentTypes: ['Linkage'] },
+    serializationStrategy: 'id',
+    target: { type: WORKFLOW_TYPE },
+  },
+  {
+    src: { field: 'target_task_id', parentTypes: ['Linkage'] },
+    serializationStrategy: 'id',
+    target: { type: TASK_TYPE },
+  },
+  {
+    src: { field: 'source_task_id', parentTypes: ['Linkage'] },
+    serializationStrategy: 'id',
+    target: { type: TASK_TYPE },
+  },
+  {
+    src: { field: 'profileId', parentTypes: [`${SETTINGS_TYPE_PREFIX}Notification`] },
+    serializationStrategy: 'id',
+    target: { type: `${SETTINGS_TYPE_PREFIX}CommunicationProfile` },
+  },
+  {
+    src: { field: 'revenueRecognitionRuleName', parentTypes: ['GETProductRatePlanChargeType'] },
+    serializationStrategy: 'name',
+    target: { type: `${SETTINGS_TYPE_PREFIX}RevenueRecognitionRule` },
+  },
+]
 
 /**
  * Convert field values into references, based on predefined rules.
@@ -152,7 +52,7 @@ export const addReferences = async (
  */
 const filter: FilterCreator = () => ({
   onFetch: async (elements: Element[]) => {
-    await addReferences(elements)
+    await referenceUtils.addReferences(elements, fieldNameToTypeMappingDefs, ['id', 'name'])
   },
 })
